auto refactor application class into separate file
